+++ test/integration_tests/LoggerAddRemoveGetTest.cpp
@@ -27,7 +27,7 @@ public:
       {
         FileSinkConfig cfg;
         cfg.set_open_mode('w');
-        cfg.set_do_fsync(true);
+        cfg.set_fsync_enabled(true);
         return cfg;
       }(),
       FileEventNotifier{});